Run database migrations for Publisher during deploy
This will run database migrations using a new task Publisher application task
definition prior to updating the ECS Service to use the new task definition.

The command 'rails db:migrate' will run in a separate task_runner cluster,
and the one-off containers will not be responsible for serving user requests.

This follows the design proposal in [proposal 005].

[proposal 005]: https://github.com/alphagov/govuk-replatforming-discovery-2020/blob/1ceebbb0319ecb3b6be35e3dedea2331638cf5a8/proposals/005-ecs-fargate-deployment.md#database-migrations-will-be-performed-before-new-code-is-deployed

The run-task task is intentionally generic so that it can run arbitrary
commands in a container with the same configuration as those serving users.

Additional commits are required to enable db migrations for other apps.

Story: https://trello.com/c/jepU0c7G/232