envoy-settings instead of task-definition module
The idea here is that instead of passing everything we need to create a
task-definition _down_ to a module that knows about the envoy config
(which should be the same in every task definition), we should pass the
envoy config _up_ and create the task_definition resource directly in
the terraform deployment.

I think this makes it a bit easier to work out what's going on, and it
means we don't have to declare as many variables.