WIP - use a NACL instead of SGs for test environment IP safelist.
The idea is to reduce complexity and noise in the security group config,
as well as reduce the chances of misconfiguration.

TODO: generate the NACL entries for cross-subnet traffic (I'm pretty
sure this has to be explicitly allowed, which somewhat takes the shine
off this idea but it might still be ok - let's see).
TODO: remove the office IP safelist stuff from the security groups.
TODO: read the office IP safelist from govuk-aws TF remote state instead
of duplicating it.