Finalise publisher deployment
A couple of bits to make it consistent with content-store, now that's
been refactored.

The worker needs to override command in the container definition, so
that needs to be configurable now. Using `[]` as the default makes this
a no-op for content-store (which doesn't provide command).

It's a good idea to set REDIS_URL as well as the other two - I've raised
a tech debt card to deal with the fact that there are two ways of
setting it.