Add build jobs for apps
This commit adds image build jobs for a number of apps, based on the existing `smokey` build job. It refactors the `build-image` step of the jobs into a separate `build-image-definition.yml` task config to avoid repeating this across all jobs. The `registry-image` resource is also updated so that it now points to repositories in AWS ECR rather than Docker Hub, requiring user credentials and the assumption of a role using those credentials to access ECR.