Upgrade Terraform version: 0.13.3 -> 0.14.6
This is the latest version of Terraform:
https://www.terraform.io/downloads.html

Applying Terraform will automatically download the new version.

The biggest change is the addition of the Terraform lock files
in each deployment.

The lock files are an addition in Terraform 0.14. They're intended
to make terraform use the same dependencies (providers, mostly aws)
when it re-runs. I assume this is to make terraform builds faster
and more reproducible.

The lockfiles are created and updated when you run `terraform init`.
They should be included in version control.

One annoying flaw in the implementation of the lockfiles is that
the checksums in the lock files only include the checksums for
the particular platforms you specify. If you don't specify a platform,
only the checksums for your platform are used. This becomes a problem
when you try to run terraform validate on another platform without
first running terraform init, since the checksums between platforms
are different you'll see an error like:

```
Error: Failed to install provider from shared cache

Error while importing hashicorp/aws v3.13.0 from the shared cache directory:
the provider cache at .terraform/providers has a copy of
registry.terraform.io/hashicorp/aws 3.13.0 that doesn't match any of the
checksums recorded in the dependency lock file.
```

In plainer language, it means that when you tried to run terraform validate,
it pulled the provider@version@your_platform, but the checksum for the
provider doesn't match the checksum in the hcl lockfile (which was generated
by a terraform init on another platform, e.g. windows and you are on a mac).

The solution is to generate a lockfile that will contain checksums for all
platforms:

```
terraform providers lock \
  -platform=windows_amd64 \ # 64-bit Windows
  -platform=darwin_amd64 \  # 64-bit macOS
  -platform=linux_amd64     # 64-bit Linux
```

This is documented here:
https://www.terraform.io/docs/cli/commands/providers/lock.html#specifying-target-platforms