Introduce www/draft-origin for assets
This PR is the continuation of the work to add Fastly CDN in front of
the Test environment for basic authentication.

In order to route all public user traffic (except traffic usually going to
assets.*.publishing.service.gov.uk) via the CDN so that they can be
authenticated, we need to introduce the origin component of GOV.UK.

The origin component of GOV.UK acts as a `single` entry point to the
GOV.UK environment and fans out the requests to different other apps of
GOV.UK. The origin component of GOV.UK is implemented in the EC2 platform
using nginx configs and GOV.UK router app

This PR implements the nginx routing using a AWS Application Load
Balancer in the new ECS environment with the use of path prefix
matching to redirect requests to the correct frontend GOV.UK apps.

In effect, if HTTP request arrives with:
1. /assets/frontend/* this is directed to the Frontend app
2. /assets/static/* this is directed to the Static app
3. otherwise, all traffic goes to the Frontend app
   (when Router is properly implemented, this should go to the Router
app)

Ref:
1. [trello card](https://trello.com/c/TjJOKSKp/404-basic-auth-for-integration-environment)