Replace 'dummy' bootstrap task definition
During bootstrapping we need to create a task definition
for a service to be brought up.

Previously we just used a dummy task definition that would
be replaced by a real task once the apps are deployed.

This bootstrap task definition was confusing - here it is
replaced with another bootstrap task def. The difference
is that this one should match the definition created later
when we deploy the app. This will enable us to bootstrap
a new environment in one step.