Provide ((workspace) to update-ecs-service
This is needed because the ECS cluster that needs to be updated is
namespaced by workspace. In the default workspace it will be
"govuk-ecs", in the towers workspace it will be "govuk-towers".

We could drop a conditional here if we called the default cluster
"govuk-default", instead of "govuk-ecs". I think that's a wider
discussion though.