Only attempt to push rubygem if local version is newer
This has been opened to resolve a scenario where errors can occur with
this action. These errors occur is if a build of the main branch is delayed
and this action runs after a new version of the gem has been created. This
happened with [1] where a delayed build of the main branch for gem version
6.8.2 ran after the release of 6.8.3.

Rather than doing a basic version equality test, this instead parses the
version so it can compare that the version is actually newer not just
different. It does this by using the Gem::Version class of Ruby stdlib.

The fact that this errored didn't cause any behavioural problems, just
the inconvenience of the alerts following a failed action.

---

In order to resolve this I did some research (and was given some
pointers) on approaches. This seems an appropriate place to document
them.

The interesting thing to consider, for a tool that does a form of
continuous deployment based off GitHub Actions, is that it isn't
guaranteed that a GitHub Action running against the main branch is
building against the HEAD of that branch. As we found in the situation
explained above, it's wholly possible to have HEAD run and then HEAD~1
run. This gets tricky if we want to deploy the code based off a
successful branch run as a later branch could mean we're actually
deploying a regression of the code that is commits before the main
branch's tip.

An idea on how this can be resolved is using the `concurrency` [2]
feature of GitHub Actions. This feature allows for there to only be one
job running at a time and to cancel any pending jobs when a new one is
triggered. I decided that using this actually risked making the
situation worse. For example in the above scenario of HEAD~1 being
kicked off after HEAD, assuming we had concurrency set at a branch
level, this would mean that HEAD could never get build as the queueing
of HEAD~1 would cancel it.

In theory the best way to deal with the possibility of this out-of-order
handling is to have a means to know that the commit being built is the
current HEAD of the branch. Some options for this would be to run the
action against the branch rather than a particular commit, which is
GitHub's default (this can be done by configuration actions/checkout to
have a branch as a ref) or to clone the branch deeper and check the
commit being built is actually the HEAD of that branch. Cloning against
the branch directly could be risky as it could be a different version of
the codebase than what is tested.

However, for this scenario I didn't want to get complex on the Git side
and decided that this new conditional was sufficient to protect against
out-of-order causing failed builds.

[1]: https://github.com/alphagov/govspeak/runs/6982747734?check_suite_focus=true
[2]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#concurrency